src/bff/pyproject.toml
[project]
name = "apic-vibe-portal-bff"
version = "0.0.0"
description = "Backend-for-Frontend API for APIC Vibe Portal"
requires-python = ">=3.14"
dependencies = [
    "cachetools>=5.5.0",
    "cryptography>=46.0.5",
    "fastapi>=0.135.3",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.0",
    "starlette>=0.46.0",
    "structlog>=25.5.0",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "pytest>=8.0.0",
    "httpx>=0.28.0",
    "pytest-asyncio>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["B008"]  # B008: Allow Depends() in function defaults (standard FastAPI pattern)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
